Fixed deadlocks if computes spawn their own threads.
We need to release the GIL on the thread which requests the compute, so that if the compute spawns threads, they can enter into Python.

The only example we have of this right now is the Instancer when getting the bounds, but I've fixed all the ValuePlug::getValue() bindings too for good measure.