Fixed SceneView bug where wrong context was used in the SceneGadget.
It was necessary to make View::setContext() virtual to allow this to happen, but that seems consistent with the direction we're taking where Views will internally use helper Gadgets that do their own drawing outside of the nasty View::update() mechanism. Also added a \todo about phasing out the latter mechanism.

Binary incompatibility :

- View::setContext() is now virtual