Added SceneTestCase.assertBuiltInSetsComplete() method.
We were already checking that all objects referenced by sets do exist - this extends upon that by checking that all cameras, lights and coordinate systems that exist are referenced by the appropriate built in sets. The test has to be disabled for a couple of tests involving Alembic files, because we don't have a way of flagging cameras in those.