Added sneaky hotkey for navigating inside any node in NodeGraph.
This is useful for seeing inside References and other nodes which have internal networks. We really need to put the NodeGraph in a read-only mode for this, but since we don't yet have one, we're consoling ourselves that the naive user is unlikely to hit upon this particular key combo.