Added CropWindowTool.
This allows the crop window to be set interactively within the 3d view when looking through the render camera. The tool operates on the first upstream Options node with an active crop window. If no such node is found, it will activate the crop window on the first upstream Options node with an inactive crop window. If no Options nodes exist, it prompts the user to insert one.