Optimised Switch node for constant indices.
When the index is not being output from a ComputeNode, we know that it doesn't vary by context and can optimise the switch with an internal connection. A benchmark with the bunny scene followed by 50 switches shows a 6x speedup overall - a 30x speedup in the Switch operation when the scene read time is factored out. That's great, but what it really means is that we need to do better in Switch::oppositePlug() for when the index does vary with the context.