Fixed ObjectSource to work after create/undo/redo.
It was using an outdated way of using the result of an internal computation in the computation of another output - since DependencyNode::affects() is now (and has been for a long time) called for output plugs as well as inputs during dirty propagation there is not need for the additional intermediate plug.

Although this fix is good in that it tidies up some code, I suspect there's another problem somewhere which caused the input to the internal plug to be removed upon undo but not restored upon redo. This may need investigating in its own right at some point.

I have no idea why the unrelated SceneWriterTest was mucking about with the private __inputSource plug - it had no business being there.