Improved StandardNodeGadget drag/drop handling.
StandardNodeGadget accepts drag enter/move/leave/drop in order to snap drags to the nearest nodule, which makes the node graph much easier to use. But it was returning true from these event handlers regardless of whether or not it was actually performing any snapping. It now returns false if not performing snapping, enabling other handlers to step in and do their own thing.