Added Dot node.
This is just a simple organisational tool for complex graphs. It has only "in" and "out" plugs which must be created using the `Dot::setup()` method after construction - since Gaffer's plugs are strongly typed, it's not possible for the Dot to create plugs upfront. Internally the input is passed through directly to the output via a connection, so the dot does nothing computationally, it's just useful for adjusting the layout of the graph.