Added PlugValueWidget.childPlugValueWidget().
Prior to this we had just been implementing the method directly on child classes, with no default implementation in the base class. This was because previously Plug could not have children (only CompoundPlug could), so it made no sense on the base class. Now that Plug can have children and CompoundPlug is flagged for removal, it makes sense to have the method defined on the base class.

Fixes #1123.