Reimplemented parts of PathListingWidget in C++ for performance.
The key component of the PathListingWidget is the QAbstractItemModel which exposes Gaffer::Path to Qt. This is now reimplemented in C++, giving much improved performance. The column classes for extracting data for display are also now implemented in C++, meaning that the drawing of the PathListingWidget now takes place without hitting Python at all. The new PathModel class also performs sorting directly without the indirection entailed with the previous QSortFilterProxyModel approach.

This is the first time we've used Qt directly from C++, and it required a somewhat tricky approach to keep Gaffer compatible with both PyQt and PySide, and to continue to do so with a single build. The problem is that we can't simply bind our PathModel directly into Python because for PyQt we'd have to use SIP and for PySide we'd have to use Shiboken. Instead we adopt an approach where we instead make a set of manual bindings using boost::python, passing the address of a QTreeView explicitly retrieved with the new GafferUI._qtAddress() method. This perhaps isn't as elegant as making a proper binding where the PathModel appears as a class in Python, but I think it's justified by the utility of keeping compatibility with both PyQt and PySide in one build, and by not maintaining two sets of bindings in systems we're not familiar with.

Now that libGafferUI links directly with Qt we may need to review our builds intended for use in hosts such as Maya, Nuke and Houdini. Whereas formerly it was sufficient just have a host-compatible PyQt build on the system, we may now need to ensure that we build libGafferUI against the same Qt version as the host uses.