Detemplatized GafferScene::ObjectSource.
This was unnecessary complexity. The idea was originally that we might want to also template using the FileSource base class to allow the creation of something using IECore::ObjectReader behind the scenes. But now FileSource doesn't exist, and it's clear that we want to deal with additional filetypes by introducing a SceneInterface for them and using the SceneReader.