Fixed SceneReader behaviour for invalid files and paths.
Previously, it would throw an exception on the first attempt, and then either silently fail or crash on subsequent attempts. This was because we were updating the fileName/path fields in LastScene _before_ the operations which might throw, rather than after. By only updating them after, they remain in sync with the true state of the corresponding fileNameScene/pathScene fields.