Improved StandardNodeGadget error display for large graphs.
When the source of the error might be far away (or at a different subgraph level) from the node being viewed, it helps to have a trail of errors reported from the viewed node all the way to the source. This is achieved by displaying an error on every intermediate node, and including a header specifying whether the error occurred internally, upstream, or on a particular plug of the node.