Fixed ComputeNodeWrapper exception handling.
It was leaking boost::python::error_already_set out into the wider C++ world, where it cannot be understood (because the pure C++ libraries don't (and shouldn't) link to boost::python or python). Worse still, because it wasn't clearing the python error indicator, PyErr_Occurred() remained true, and when later reentering python code the error resurfaced in the wrong place (as soon as something else checked for PyErr_Occurred()).

The fix is to use translatePythonException() to both turn the exception into something the pure C++ libraries can deal with, and also to clear the python error indicator. This should be our approach to exception handling at all boundaries between pure C++ and Python.