Fixed evaluation time of look-through cameras in the Viewer.
If either transformBlur or deformationBlur was on in the globals, the camera transform was evaluated at the start of the shutter range, rather than at the current time. Since everything else in the viewer is evaluated at the current time, this led to an unwanted offset (and prevented look-through working at all with IECoreMaya::LiveScene inputs).