Added Gadget::requestRender() method.
This takes care of emitting renderRequestSignal() for all ancestors as necessary, and should be called by all Gadget derived classes whenever they change and need rerendering. This replaces the old approach of using signal handling, which is more costly in terms of memory and time. In future, we are now in a position to only construct m_renderRequestSignal when accessed externally, and to only emit it once before another render is performed.