Fixed CompoundEditor lifetime bug
It looks like the fact that __serialise function that gets declared inside CompoundEditor.__repr__ was leading to circular references. The fact that __serialise was recursive and used "self" was sufficient to set up a circular dependency in the guts of python that meant that as soon as someone called CompoundEditor.__repr__, the CompoundEditor became IMMORTAL. Well, until someone called the garbage collector. This was giving me grief when switching scenes in Caribou.