Fixed dirty propagation for Merge node.
It wasn't propagating dirtiness for the channelData plugs. I've fixed this in the FilterProcessor base class rather than the Merge derived class, because the implementation of FilterProcessor::hashChannelData() implies that all subclasses must use all input channel data. I'm not sure if that's actually a useful base class behaviour, but I thought it best to at least have the affects() method consistent with the hash() method regardless.