Checking child names in interactive render
In this commit I'm checking for changing child names when the childNames plug gets dirtied (I'm reluctant to do it during a regular update because it means extra hash/plug evaluations, sorting lists etc during something I want to be fast). If a location's child names change, the lists get compared, and locations in the scene graph that are no longer present get an m_locationPresent flag that I've added set to false. This stops the SceneGraphIteratorFilter traversing into them so they don't get updated.