Improved plug default value handling for Reference nodes.
References are often used for sharing assets in collaborative workflows. A common scenario is :

1). Artist A builds a box containing some lookdev, promotes some plugs, sets their values, and exports the box as a reference.
2). Artist B imports the reference into their script and changes some (but not all) promoted values to suit the current shot.
3). Artist A then creates a new version of the lookdev, in which one of the promoted plugs has been tweaked to improve the default look.
4). Artist B updates to the new reference.

What should happen? Prior to this commit, nothing about the promoted plugs would change in step 4 - Artist B's value would trump everything, even the values they hadn't explicitly changed. Not very useful.

This commit changes this such that :

- If artist B has modified a value, they will keep that value. This applies regardless of whether or not artist A provided an updated value.
- If artist B has not modified a value, then that value will be updated with any changes from artist A.

This is achieved by creating new default values for promoted plugs being exported from a Box - the export receives default values set from the current values at the time of export. This also moves us one step further along the road to having Boxes be a good basis for the definition of Doohickeys or Gizmos or Gaffes - because this is effectively defining a new node type, it will be very important for newly constructed nodes to be at their default values.

This fixes Image Engine internal ticket #7109.