Fixed ReferenceTest.testDefaultValueClashes().
It was written at a time before we reset default values when exporting boxes for referencing, and falls foul of the fact that the default value on the reference is not the same as the default value on the box. The important thing is that the plug values themselves stay the same, and that is now what the test case asserts.