Fixed creation of expressions for BoolPlugs.
Because there was no value being assigned in the default expression, parsing failed, and the expression editor was not shown. Now we always use the current plug value for the default value of the expression, which also means that the initial value for the expression will always be valid (otherwise it could be out of range for instance).