Fixed context used by scene view camera chooser.
The SceneViewUI creates a ScenePath for the path chooser using the context stored on the SceneView, so we must ensure that the context is set correctly before creating the toolbar.

This is something of a stopgap fix. Every editor (including the Viewer) has a `setContext()` method which can in theory be used to allow different editors to be viewing different times (or different proxy levels, or different passes, or anything else that you want to represent within a context). The right fix would be to override `Viewer.setContext()` to set the context on each of the toolbars, and have that passed down through the PlugLayout to the `PlugValueWidget.setContext()` methods, and then to use a new ScenePathValueWidget which overrides `setContext()` to update the context on the ScenePath used by the path chooser. But neither NodeUI, nor NodeToolbar, nor PlugLayout currently provide a `setContext()` method, so there's a gap in the chain. Fortunately, none of the `setContext()` functionality is exposed in the current UI, so in practice the context is always the default one held in `ScriptNode::context()`, and this fix suffices for now.