Added -threads command line argument to Gaffer.Application.
This uses a tbb::task_scheduler_init() object to set the maximum number of threads that tbb will use. It would still be possible for an application to use more threads by spawning its own threads explicitly, but we should avoid doing that for heavy computation via the node graph.

Note that when running applications in a host (Maya for instance), we return from run() immediately. In this case, the task_scheduler_init object will be exited immediately too. My understanding is that in this case, any tbb configuration done by the host will win out over the brief configuration we did. If the host does no configuration, then I believe we end up back using the default (all cores). For this reason, the -threads argument is best considered for use only when running Gaffer standalone.