Added exception to highlight unimplemented Path.copy() methods.
The old Python based Path base class had a default implementation for copy() that used __class__ to ensure that the copy had the right type. That's not possible in the new C++ base class. We could reimplement the behaviour via the bindings, but experience with the old implementation shows that the default implementation was rarely useful, because most subclasses have their own private data which also needs copying. It seems better to highlight missing implementations rather than have harder to diagnose failures coming from relying on an incomplete default implementation.