Removed GraphComponent info() implementation.
It was broken anyway - since __dictEntry is not defined it always failed to provide "graphComponent:type". We might want to introduce some properties in the future, but until we need them it's good to have one more test case in the transition from info to properties - a derived class that overrides neither set of methods.