Fixed ScenePath crash when target plug is removed from its node.
In this case, we were failing to disconnect from Node::plugDirtiedSignal() when the ScenePath was destroyed. Subsequent emissions of that signal would then call ScenePath::plugDirtied() on the dead ScenePath - boom.

Alternative (and more automatic) solutions would be :

- Deriving Path from boost::signals::trackable
- Using boost::signals::scoped_connection to store the connection

Each of those has more memory overhead than the `NodePtr m_node` solution though, and since we currently make large numbers of Paths in the PathListingWidget, the more frugal option seems best.