Reduced overhead of SceneInspector "inherited" indicator.
The SideBySideDiff makes the corner widget visible for each value in update(). Then the DiffRow was hiding it again for non-inherited attributes. This on/off visibility toggle caused a lot of unnecessary computation in Qt. In contrast, changing the label text is much cheaper, and QLabel has a short-circuit for the case where the text is not changing.

This doubles the speed of update for the SceneInspector when viewing 100 attributes.