Added node graph bookmarks.
Nodes may be bookmarked using the right click context menu in the node graph. They are then available in the right click context menu for plugs in the node graph, allowing connections to be made easily to distant nodes. I suspect we'll want to expand on this to give access to bookmarked nodes elsewhere in the UI at a later date, but for now it should be sufficient for #849, as discussed on #413.

Fixes #849.