Renamed GafferScene::Filter "match" plug to "out".
This brings it into line with other nodes with a single primary output, like SceneNode and ImageNode. It also paves the way for producing a FilterSwitch node using the Switch mixin class.

Backwards compatibility is provided by an overridden __getattr__ implementation for the Filter class - note though that any comparison of the plug name will no longer work, as demonstrated in the necessary changes to the unit tests and the UI files.