Fixed failing ImageWriter test.
I have no idea why this is passing on the Travis builds, but failing at Image Engine. But it seems to be failing back to Gaffer 0.8.0.0 at Image Engine, so I'm fairly confident it wasn't introduced by this branch. In any case, the change doesn't affect the integrity of the test, so I'm going to leave that mystery for another time...