Moved propagateDirtiness() from DependencyNode to Plug.
The main motivation for this is that an upcoming commit will require propagateDirtiness() to have access to a protected method of Plug. Rather than declare two-way friendship between DependencyNode and Plug, it seems more appropriate to take a route requiring no friendship declarations at all.

This reorganisation also makes for nice parallels between the DependencyNode/Plug and ComputeNode/ValuePlug relationships. In both cases now, the node provides methods that define only internal relationships (affects(), hash() and compute()) between plugs, and then plugs themselves are responsible for the external graph relationships defined by connections, and for triggering the methods provided by the nodes.

This was a private method, so API compatibility is not affected in any way.