Reversed direction of graph used for dirty propagation.
We were using reverse iterators everywhere to work around the fact that topological_sort() returned the results in reverse order. It is more straightforward to simply build the graph with edges in the other direction. This also happens to be essential to get the correct ordering for the batched dirty propagation to be implemented shortly.