Fixed environment variable expansion to allow persistent hash cache.
It's a fundamental rule of Gaffer's computation engine that hashes and computation results must not depend on anything other than the state of the graph (plug values and connections) and the contents of the Context. Allowing environment variable substitutions in Context::substitute() and then using them in StringPlug::hash() and StringPlug::getValue() violates this rule, and perhaps was unwise. Now that ValuePlug caches hashes, it has no way of knowing when to invalidate the cache based on changes to environment variable, and changing an environment variable can poison the hash.

We work around this by taking a copy of the environment at startup, and using this to perform the substitutions. This also has the added benefit of improved performance - testManyEnvironmentSubstitutions() shows 12% improvement with a standard CentOS bash shell environment, and 62% improvement when 1000 additional environment variables are added.