Labels for plugs set to their userDefault no longer display as changed.
Note that for plugs with userDefaults, the labels now display as changed when set to the real default value. Some people seemed confused that there is any difference between the real default and the userDefault. After using it a bit, it seemed more natural that the user-facing UI should consider the userDefault as the true default, even if the API doesn't.

Fixes #1216