Fixed crash bug when deleting nodes in plugInputChanged().
When the input is changed, we propagate dirtiness and the plugs end up in the DirtyPlugs graph. Then the nodes get deleted and the DirtyPlugs graph becomes the sole owner of the plugs. Then it clears the graph, the plugs are deleted, and they trigger an new dirty propagation as their children are unparented. Boom. To work around this we use a flag to avoid this reentrant clear/emit behaviour. A better long term solution might be to add support for weak pointers to RefCounted, and use those within DirtyPlugs.

This also required a fix for an egregious bug in the ScopedAssignment class.