OpDialogue sets userDefaults for all plugs during initialization.
This allows the initial parameter values to show as default in the UI, while still allowing interactive user changes to be visibly distinguishable. Note that we're only doing this when the OpDialogue is initialized with an Op instance, and not when it is given an OpHolder directly, on the assumption that a client providing a node directly has already configured it as desired.