Reimplemented PathMatcher::Iterator as RawIterator and Iterator.
The RawIterator visits every path in the tree, and the Iterator is effectively a filtered version which skips all non-terminal locations. The two iterators are necessary, because some algorithms (Prune set computation for one) can be optimised by pruning iteration as non-terminal locations. Separating out the logic of iteration (including the nasty case of the root) and filtering non-terminal locations also makes for clearer code, and fixes the previously failing test cases.