Optimised Prune set computation.
We now use the new PathMatcher::RawIterator to avoid the expense of calling PathMatcher::paths(),
and because we're now performing a recursive iteration including the ancestor paths, we can prune the iteration and avoid filter lookups in many cases. This gives a 20% - 70% reduction in compute time when pruning a large (0.5 million paths) production set.