Added unit test for threaded computation while editing another graph.
This exposes a bug which is also exposed by GafferScene.SceneNode.testCacheThreadSafety(), but this time in a much more straightforward way. The SceneNode test only fails when run after other tests, when the python garbage collector decides to destroy some old nodes from the previous test, in the middle of the threaded computation performed by testCacheThreadSafety(). The new test is pure C++, doesn't rely on unlucky garbage collection timing, and reproduces the problem even when run alone.