Fixed threading test failures.
These occurred in unit tests which concurrently modified one graph while performing threaded computations on another. This isn't a pattern we currently use in practice, but it's something we want to remain possible. And as SceneNodeTest.testCacheThreadSafety() showed, its possible to do it inadvertently if python's garbage collector delays destruction of an old graph till such a time as you're computing with a new one.