Added metadata presets support for OSL color and point parameters.
Previously this was not possible because PlugValueWidgets for the CompoundNumericPlugs would be updated half way through a setValue() call, when not all the child plugs were in their final state. Now that we properly batch operations for dirty propagation, the UI only updates when all operations are complete, and the plug is in a consistent state.