Path matcher string comparison fix
It turns out that PathMatcher::Node::wildcardsBegin() sometimes skips out all the wildcards because all interned strings happen to be < the empty interned string. This depends on the empty string's memory address, and was giving inconsistent behavior in renders (it never seems to be break things in single threaded mode for some reason). I'm not really sure how to write a test case for this - any ideas? Anyway, I've fixed it by modifying PathMatcher::Name::operator<() so all InternedStrings are > the empty string.