Fixed PathMatcher wildcard matching bug.
PathMatcher::Node::wildcardsBegin() was sometimes returning the wrong value, and we were therefore skipping matches. The logic for wildcardsBegin() was assuming that InternedString( "" ) would always compare less than InternedString( "anything" ), but this is not true at all. It would be true if we compared the string contents, but we compare the string addresses for speed. We were just getting lucky and the tests were passing because InternedString( "" ) is constructed during IECore initialisation, and would typically get a lower address than anything else. In the wild though, this is most certainly not the case. The hard work in figuring all this out was done by Mr David Minor.

To fix this, we replace the hasWildcards bool with an enum containing an explicit intermediate Boundary value.