Fixed test failure
GafferRenderManTest.InteractiveRenderManRenderTest.testMoveCoordinateSystem seems to have been checking image pixels too close to the edge of the image, and reading pixels that were a bit more black than expected because of this (the images were correct from visual inspection). I've moved the sample positions away from the edges of the image to avoid this problem.