PlugLayout : Added support for negative "layout:index" values.
These work in the same way as negative indices in python - indexing backwards from the end. This will be useful because often the first plugs on a node aren't of primary interest to the user (DependencyNode "enabled" plugs for instance), and we'd like to place them on a later tab.