ImageWriter : Fixed TIFF pixel aspect ratio bug.
When adding metadata to the OIIO spec, we were always appending a new value, when we should have been editing an existing one in the case the field was already there. It appears that meant we ended up with two "PixelAspectRatio" fields - the one from the Gaffer image metadata, and the one we specified explicitly from the format. Because the explicit one was second, it was ignored, and we wrote the wrong pixel aspect into the file. This manifested itself as a GafferImageTest.ImageWriterTest.testTiffMetadata() failure with OIIO 1.5.

Since the ImageSpec::attribute() utility method already does the work of searching for an existing entry, we just use that.