Deprecated Plug::PerformsSubstitutions flag.
It was always awkward that this was defined on the base class but only made any sense on the derived StringPlug class. Furthermore it now provides only very crude control compared to the new StringPlug constructor argument.

Replaced all uses of the flag with the new method, except GafferCortex::TypedParameterHandler - this will require a slightly more complex change (adding a FileSequenceParameterHandler) and will be best addressed when we also address automatic variable substitution for ExecutableNodes (#887). In the meantime, it provides test coverage for the operation of the deprecated feature.