FilterPlugValueWidget : Updated for StandardNodeUI changes.
Because we can no longer request a simplified UI, we just embed a complete UI for now. I'm not sure how I feel about this - on the plus side Filter nodes regain control of their layout, and on the negative side the extra tabs might seem a bit cluttered. It'll be good to get some user feedback on this. The plan for restoring the old behaviour would be to embed a PlugLayout, and add some stuff to PlugLayout to allow sections to be filtered/renamed from the outside. I do think it's best if PlugLayout is solely in charge of its internals though, so I'm not super keen on this.