ParentConstraint : Stopped discarding local transform.
Prior to this we just made the new full transform of the constrained object match the full transform of the target. This meant that using a ParentConstraint was not directly equivalent to using a Parent node - the user was expected to put the original local transform into the "relativeTransform" plug manually if they wished to include the original local transform and get the exact same effect that parenting would have had. This was based somewhat on Maya's parent constraint behaviour where if "Maintain offset" is used, the necessary offset is stored on the constraint node and applied during computation. This sucked for several reasons :

- We weren't automatically filling in the "relativeTransform" plug values to maintain any offset. This was something the user had to do.
- There is one "relativeTransform" plug but the constraint can operate on many objects specified by a filter. So it wasn't even possible to fill it in appropriately for multiple objects.
- "Maintain offset" doesn't make sense in a procedural system like Gaffer - because we're not manipulating a single static hierarchy, there's no concept of "the scene state in which the user created the constraint", so there's no sensible point at which to automatically stash the offset.

But because we're a procedural system, we still have access to the input local transform in every constraint compute, and can emulate the true effect of parenting without resorting to stashing relative offsets anywhere. So we simply use that and our ParentConstraint is now directly equivalent to the Parent node in its effect on the full transform of an object.

In short, we've gone from "half-assed emulation of what Maya calls a parent constraint and isn't very suited to use in Gaffer", to "directly equivalent to actual parenting, and dead easy to implement and use in Gaffer". Thanks is due to Daniel Dresser for making the case for this.

Backwards incompatibility :

- Changed ParentConstraint behaviour to include the local transform of the constrained object.