Removed TransformPlugValueWidget.
This derived from CompoundPlugValueWidget, which we are going to remove once we've completed the transition to using LayoutPlugValueWidget in preference. Although TransformPlugValueWidget did provide a handy summary heading, in practice this wasn't used because we always placed our transform plugs on tabs of their own, so we're not losing anything.

A search on GitHub shows we're not using this directly in any projects, but the UI for internal IE VDB nodes will change as a result - we need to move those transform plugs to their own tabs anyway though, to match the standard Gaffer nodes.

Incompatibility :

- Removed TransformPlugValueWidget.