Box : Removed asUserPlug arguments from promotion methods.
Promoted plugs will now always be parented directly to the Box node, freeing up the "user" plug for its usual purpose. This is important to allow the "user" plug to be used freely on Reference nodes, and will be even more important when we start exporting Boxes as new node types (GiftBoxes or Doohickeys or whatever).