PlugLayout : Fixed ObjectWriterUI.
The ObjectWriter is very naughty and registers a ParameterValueWidget in place of a PlugValueWidget. Really we need to refactor GafferCortexUI so that this is not necessary, but in the meantime we make PlugLayout tolerant of the behaviour in the same way that StandardNodeUI was. This fixes a bug introduced by 0cec503cf5e1c868a9b74d3cec42d87b97f001e0.