Apps : Fixed errors caused by special characters in .gfr filenames.
Several of the apps tried to name the ScriptNode based on the filename, resulting in errors because filenames may contain characters not accepted as GraphComponent names. Since our apps (and the FileMenu) were inconsistent in the names they were applying anyway (and there was no syncing of the node name on subsequent filename changes), we now simply just use the default name for the ScriptNode.