UIEditor : Refactored to support sectioning.
The plug listing column now also shows the hierarchy of sections which plugs belong to, and plugs and sections can be reparented and reordered with drag and drop. Also factored out the panel which edits the plug name and description into its own class, since it will be growing in the future.

Backwards incompatibilities :

- Replaced UIEditor setSelectedPlug()/getSelectedPlug() methods with setSelection()/getSelection() methods. These allow sections to be selected in addition to plugs.