PathFilterUI : Added support for dropping objects into NodeGraph.
Objects can be dragged to either a PathFilter, a FilteredSceneProcessor with a PathFilter attached, or a FilteredSceneProcessor with no filter (in which case one will be created). Shift+Drag adds objects to the filter, Control+Drag removes them, and a plain drag replaces the current value.

Unfortunately, because we must store the connection returned by `connect()` to keep it alive, we currently have to leak connections by keeping them forever in our __connections list.