NodeGraph : Improved "Select Affected Objects" menu item.
Previously this menu item had only been available on FilteredSceneProcessor nodes, because that is the one place where an input scene and a filter are unambiguously associated. But users would often be looking for the menu item on the Filter nodes themselves. This makes the menu available there too, selecting the union of all objects affected by the filter, on all FilteredSceneProcessors which it is attached to. We recurse down the output connections of the filter to find such FilteredSceneProcessors, so that we can discover them inside Boxes and on the other side of Dots etc.