Signal binding : Added "scoped" argument to connect() method.
Which isn't as simple as it sounds, because we've also simplified the bindings for the connection classes, and removed the tangle between connection/signal/slot bindings.

The signal bindings have had a slightly complex history, and they're unfortunately an example of where we've deviated slightly from our "the python API should mirror the C++ API as closely as possible" rule. In C++ when you connect a slot, the signal keeps a reference to the slot, and a boost::signals::connection is returned which provides a handle to this signal/slot association. You're not required to keep a hold of the connection, so if you don't assign it to a variable, the actual connection still exists. In C++ we can also derive from boost::signals::trackable, which provides automatic connection management that disconnects the slot automatically if it's a member function of a trackable object which has died. Alternatively, you can assign the new connection object to a scoped_connection member variable, and the connection will be removed when that goes out of scope or is reassigned.

In Python however, we can't derive from trackable, or at least it's a concept that we've never explored. We also had big problems early on with circular references between widgets, their signals, and slots which were methods of the widget - meaning that widgets could live forever. We originally dealt with this in a very complex manner, with connect() returning a custom connection binding which acted more like a scoped_connection, but held the slot itself internally in __dict__. This along with a patched boost::python allowed these circular connections to eventually be garbage collected. But we found that widgets-awaiting-collection caused their own problems and the boost patch was unreliable, invented WeakMethod instead and backed out of some of the weirdness.

As an artifact of that history, we still had a weird custom connection binding, and the slot was still stored in the connection. This commit undoes that, meaning the connection bindings are now just a simple one-to-one binding of the boost classes. This also allows us to move all the SlotCaller logic into the signal bindings, where they make more sense alongside the SignalCaller logic.

That then allows us to add the "scoped" argument to `connect()`, and return either a connection, or a scoped_connection. Passing `scoped = False` allows a style of connection management that was previously impossible in Python -if we just want the lifetime of the connection to match the slot lifetime, we no longer have to worry about finding a suitable place to stash a scoped_connection, and then find an appropriate time to remove it.

Unfortunately we still return a scoped connection by default, whereas in C++ you get an unscoped one by default. I can't see an easy way of reversing that without a huge upheaval in all code using GafferUI. I don't think keeping things the way they are is _too_ bad either, because the lack of `trackable` in python means that the scoped style is by far the most predominant one needed in Python.